update35-Math.random round & int

+++ src/main/java/Main.java
@@ -5,7 +5,7 @@ import java.util.Date;
 
 public class Main {
     public static void main(String[] args) {
-        double result = Math.round(Math.random() * 100);
+        int result = (int) Math.round(Math.random() * 100);
         System.out.println(result);
     }
 }
